+++ maverick.graph.feature.navigation/pom.xml
@@ -30,11 +30,7 @@
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.apicatalog</groupId>
-            <artifactId>titanium-json-ld</artifactId>
-            <version>1.3.3</version>
-        </dependency>
+
         <dependency>
             <groupId>org.av360.maverick.graph</groupId>
             <artifactId>graph-api</artifactId>
