+++ maverick.graph.feature.navigation/src/main/resources/context/list.jsonld
@@ -2,16 +2,12 @@
   "@context": {
     "hydra": "http://www.w3.org/ns/hydra/core#",
     "sdo": "https://schema.org/",
-    "member": { "@id": "hydra:member", "@type": "@id" },
-    "view": { "@id": "hydra:view", "@type": "@id" },
-    "limit": { "@id": "hydra:limit", "@type": "@id" },
-    "next": { "@id": "hydra:next", "@type": "@id" },
-    "previous": { "@id": "hydra:previous", "@type": "@id" },
-    "first": { "@id": "hydra:first", "@type": "@id" },
-    "last": { "@id": "hydra:last", "@type": "@id" },
-
-    "view": "hydra:view"
-
+    "hydra:view": { "@id": "hydra:view", "@type": "@id" },
+    "hydra:limit": { "@id": "hydra:limit", "@type": "@id" },
+    "hydra:next": { "@id": "hydra:next", "@type": "@id" },
+    "hydra:previous": { "@id": "hydra:previous", "@type": "@id" },
+    "hydra:first": { "@id": "hydra:first", "@type": "@id" },
+    "hydra:last": { "@id": "hydra:last", "@type": "@id" }
   },
   "@type": "hydra:Collection",
   "hydra:totalItems": {},
