+++ pom.xml
@@ -48,6 +48,7 @@
         <logback.version>1.4.11</logback.version>
         <rdf4j.version>4.3.8</rdf4j.version>
         <springdoc.version>2.2.0</springdoc.version>
+        <titanium.version>1.3.3</titanium.version>
         <swagger-annotations-jakarta.version>2.2.15</swagger-annotations-jakarta.version>
     </properties>
 
@@ -62,6 +63,11 @@
     <dependencyManagement>
 
         <dependencies>
+            <dependency>
+                <groupId>com.apicatalog</groupId>
+                <artifactId>titanium-json-ld</artifactId>
+                <version>${titanium.version}</version>
+            </dependency>
             <dependency>
                 <groupId>com.tngtech.archunit</groupId>
                 <artifactId>archunit</artifactId>
